refactor(detail): move the recording page's components and icons out
2,360 -> 1,862 lines. Six components (RetranscribeModal, SegmentEditModal,
RecordingNameForm, SpeakerRow, NoteRow, SegmentRow) and the twelve icon
constants move to components/detail, where this page's components already live.

A byte-for-byte move: 514 deletions against 16 insertions in the page, which is
the check that nothing was rewritten on the way. All 110 tests pass with one
edit - repointing the speakers test at SpeakerRow's new address, which is where
it lives now, not a change to what it does.

Two things the plan had wrong, corrected in the plan as well as the code:

There are six components, not five. SpeakerRow is declared `export function`, so
a scan matching only `^function ` misses it - and silently swallowed its 149
lines into RecordingNameForm, whose 34-line rename form suddenly needed speaker
types and play/pause icons. That mismatch is what surfaced it.

The icons had to move in the same commit rather than the next one, because
SpeakerRow renders four of them; splitting the two would have left a component
importing icons back out of the page it just left.

`fmt` came along to SegmentRow, its only remaining caller once the rows moved.
