+++ apps/web/src/components/detail/RecordingNameForm.tsx
@@ -1,0 +1,37 @@
+import { useState } from "react";
+import { useTranslation } from "react-i18next";
+
+export default function RecordingNameForm({
+  initial,
+  onSave,
+  onCancel,
+}: {
+  initial: string;
+  onSave: (name: string) => void;
+  onCancel: () => void;
+}) {
+  const { t } = useTranslation("workspace");
+  const [value, setValue] = useState(initial);
+  return (
+    <form
+      className="flex items-center gap-2"
+      onSubmit={(e) => {
+        e.preventDefault();
+        onSave(value);
+      }}
+    >
+      <input
+        autoFocus
+        value={value}
+        onChange={(e) => setValue(e.target.value)}
+        onKeyDown={(e) => e.key === "Escape" && onCancel()}
+        placeholder={t("recordingNamePlaceholder")}
+        aria-label={t("recordingNamePlaceholder")}
+        className="w-64 rounded border px-2 py-1 text-base dark:border-gray-700 dark:bg-gray-800 dark:text-gray-100"
+      />
+      <button type="submit" className="rounded border px-2 py-1 text-sm hover:bg-gray-50 dark:border-gray-700 dark:text-gray-200 dark:hover:bg-gray-800">
+        {t("common:save")}
+      </button>
+    </form>
+  );
+}
