+++ apps/web/src/pages/RecordingDetail.speakers.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
 import { describe, expect, it, vi, beforeEach } from "vitest";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { SpeakerRow } from "./RecordingDetail";
+import SpeakerRow from "../components/detail/SpeakerRow";
 import { api } from "../lib/api";
 import type { SpeakerInfo, Person } from "../lib/types";
