+++ apps/web/src/pages/RecordingDetail.tsx
@@ -37,7 +37,7 @@ import FolderChips from "../components/detail/FolderChips";
 import DownloadTranscriptModal from "../components/DownloadTranscriptModal";
 import PeopleModal from "../components/PeopleModal";
 import EditPersonModal from "../components/EditPersonModal";
-import SpeakerContactCard, { contactSummary } from "../components/SpeakerContactCard";
+import SpeakerContactCard from "../components/SpeakerContactCard";
 import SummaryEditModal from "../components/SummaryEditModal";
 import MeetingMinutesEditModal from "../components/MeetingMinutesEditModal";
 import EmailMinutesModal from "../components/EmailMinutesModal";
@@ -52,8 +52,7 @@ import { useAuth } from "../auth";
 import AttachmentsManager from "../components/AttachmentsManager";
 import CalendarLinkModal from "../components/CalendarLinkModal";
 import PreferencesModal from "../components/PreferencesModal";
-import SpeakerAssign from "../components/SpeakerAssign";
-import ToolbarButton, { iconProps } from "../components/ToolbarButton";
+import ToolbarButton from "../components/ToolbarButton";
 import FormulasToolbar from "../components/FormulasToolbar";
 import FormulasPanel from "../components/FormulasPanel";
 import FormulaRunModal from "../components/FormulaRunModal";
@@ -64,38 +63,21 @@ import { copyRichLink, transcriptUrl } from "../lib/clipboard";
 import { segmentIndexAtMs, parseMatchTimes } from "../lib/transcriptNav";
 import { speakerRanges, selectedRanges, rangeAt, nextRangeStart, type PlayRange } from "../lib/segmentPlayback";
 import { formatBytes, formatDate, formatDuration, formatDurationApprox } from "../lib/format";
-import { hasRevisions, segmentText } from "../lib/transcriptView";
+import { hasRevisions } from "../lib/transcriptView";
 import { fetchLanguages } from "../lib/languages";
 import { selectedMeetingType } from "../lib/meetingTypes";
-import type { MeetingNote, SegmentDto, SpeakerInfo, FormulaResult } from "../lib/types";
+import type { SegmentDto, FormulaResult } from "../lib/types";
+import RetranscribeModal from "../components/detail/RetranscribeModal";
+import SegmentEditModal from "../components/detail/SegmentEditModal";
+import RecordingNameForm from "../components/detail/RecordingNameForm";
+import SpeakerRow from "../components/detail/SpeakerRow";
+import SegmentRow, { NoteRow, type SegmentAssign } from "../components/detail/SegmentRow";
+import {
+  RefreshIcon, PencilIcon, MailIcon, UsersIcon, SlidersIcon, PlayIcon,
+  PauseIcon, SelectIcon, MergeIcon, TrashIcon, GlobeIcon, EyeIcon,
+} from "../components/detail/icons";
 
-// Feather-style icons for the panel toolbars and the per-speaker play control.
-const RefreshIcon = (
-  <svg {...iconProps}><polyline points="23 4 23 10 17 10" /><polyline points="1 20 1 14 7 14" /><path d="M3.51 9a9 9 0 0 1 14.85-3.36L23 10M1 14l4.64 4.36A9 9 0 0 0 20.49 15" /></svg>
-);
-const PencilIcon = (
-  <svg {...iconProps}><path d="M11 4H4a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7" /><path d="M18.5 2.5a2.12 2.12 0 0 1 3 3L12 15l-4 1 1-4 9.5-9.5z" /></svg>
-);
-const MailIcon = (
-  <svg {...iconProps}><path d="M4 4h16a2 2 0 0 1 2 2v12a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V6a2 2 0 0 1 2-2z" /><polyline points="22,6 12,13 2,6" /></svg>
-);
-const UsersIcon = (
-  <svg {...iconProps}><path d="M17 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2" /><circle cx="9" cy="7" r="4" /><path d="M23 21v-2a4 4 0 0 0-3-3.87" /><path d="M16 3.13a4 4 0 0 1 0 7.75" /></svg>
-);
-const SlidersIcon = <svg {...iconProps}><line x1="4" y1="21" x2="4" y2="14" /><line x1="4" y1="10" x2="4" y2="3" /><line x1="12" y1="21" x2="12" y2="12" /><line x1="12" y1="8" x2="12" y2="3" /><line x1="20" y1="21" x2="20" y2="16" /><line x1="20" y1="12" x2="20" y2="3" /><line x1="1" y1="14" x2="7" y2="14" /><line x1="9" y1="8" x2="15" y2="8" /><line x1="17" y1="16" x2="23" y2="16" /></svg>;
-const PlayIcon = <svg {...iconProps}><polygon points="5 3 19 12 5 21 5 3" /></svg>;
-const PauseIcon = <svg {...iconProps}><rect x="6" y="4" width="4" height="16" /><rect x="14" y="4" width="4" height="16" /></svg>;
-// Play-all = a "from the start" glyph (skip-to-start bar + triangle).
-const SelectIcon = <svg {...iconProps}><path d="M9 11l3 3L22 4" /><path d="M21 12v7a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h11" /></svg>;
-const MergeIcon = <svg {...iconProps}><path d="M6 3v6a6 6 0 0 0 6 6 6 6 0 0 0 6-6V3" /><line x1="12" y1="15" x2="12" y2="21" /></svg>;
-const TrashIcon = <svg {...iconProps}><polyline points="3 6 5 6 21 6" /><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" /></svg>;
-const GlobeIcon = <svg {...iconProps}><circle cx="12" cy="12" r="10" /><line x1="2" y1="12" x2="22" y2="12" /><path d="M12 2a15.3 15.3 0 0 1 4 10 15.3 15.3 0 0 1-4 10 15.3 15.3 0 0 1-4-10 15.3 15.3 0 0 1 4-10z" /></svg>;
-const EyeIcon = <svg {...iconProps}><path d="M1 12s4-7 11-7 11 7 11 7-4 7-11 7-11-7-11-7z" /><circle cx="12" cy="12" r="3" /></svg>;
 
-function fmt(ms: number): string {
-  const s = Math.floor(ms / 1000);
-  return `${String(Math.floor(s / 60)).padStart(2, "0")}:${String(s % 60).padStart(2, "0")}`;
-}
 
 export default function RecordingDetail() {
   const { t, i18n } = useTranslation();
@@ -1876,485 +1858,5 @@ export default function RecordingDetail() {
         />
       )}
     </div>
-  );
-}
-
-/// Asked for when the user re-transcribes: optional diarization speaker-count hints (the exception, not
-/// the norm — used mainly to split two people the diarizer merged into one).
-function RetranscribeModal({
-  initialMin,
-  initialMax,
-  hasRevisions,
-  busy,
-  onCancel,
-  onConfirm,
-}: {
-  initialMin: number | null;
-  initialMax: number | null;
-  hasRevisions: boolean;
-  busy: boolean;
-  onCancel: () => void;
-  onConfirm: (min: number | null, max: number | null) => void;
-}) {
-  const { t } = useTranslation("workspace");
-  const [min, setMin] = useState(initialMin != null ? String(initialMin) : "");
-  const [max, setMax] = useState(initialMax != null ? String(initialMax) : "");
-
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => e.key === "Escape" && onCancel();
-    document.addEventListener("keydown", onKey);
-    return () => document.removeEventListener("keydown", onKey);
-  }, [onCancel]);
-
-  function confirm() {
-    onConfirm(min.trim() ? Number(min) : null, max.trim() ? Number(max) : null);
-  }
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/40 p-4" onClick={onCancel}>
-      <div
-        role="dialog"
-        aria-label={t("retranscribeTitle")}
-        className="w-full max-w-md rounded-lg border bg-white p-5 shadow-xl dark:border-gray-700 dark:bg-gray-900"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <h2 className="mb-1 text-base font-semibold dark:text-gray-100">{t("retranscribeTitle")}</h2>
-        <p className="mb-4 text-xs text-gray-500 dark:text-gray-400">{t("retranscribeHelp")}</p>
-        {hasRevisions && (
-          <p className="mb-4 rounded border border-amber-300 bg-amber-50 px-3 py-2 text-xs text-amber-800 dark:border-amber-700/60 dark:bg-amber-900/30 dark:text-amber-300">
-            {t("retranscribeRevisionsWarning")}
-          </p>
-        )}
-        <div className="flex items-center gap-4 text-sm">
-          <label className="flex items-center gap-2 text-gray-600 dark:text-gray-300">
-            {t("minSpeakers")}
-            <input
-              type="number"
-              min={1}
-              value={min}
-              onChange={(e) => setMin(e.target.value)}
-              aria-label={t("minSpeakersAria")}
-              className="w-20 rounded border px-2 py-1 text-sm dark:border-gray-700 dark:bg-gray-800 dark:text-gray-100"
-            />
-          </label>
-          <label className="flex items-center gap-2 text-gray-600 dark:text-gray-300">
-            {t("maxSpeakers")}
-            <input
-              type="number"
-              min={1}
-              value={max}
-              onChange={(e) => setMax(e.target.value)}
-              aria-label={t("maxSpeakersAria")}
-              className="w-20 rounded border px-2 py-1 text-sm dark:border-gray-700 dark:bg-gray-800 dark:text-gray-100"
-            />
-          </label>
-        </div>
-        <div className="mt-5 flex justify-end gap-2">
-          <button
-            type="button"
-            onClick={onCancel}
-            className="rounded border px-3 py-1.5 text-sm hover:bg-gray-50 dark:border-gray-700 dark:text-gray-200 dark:hover:bg-gray-800"
-          >
-            {t("common:cancel")}
-          </button>
-          <button
-            type="button"
-            onClick={confirm}
-            disabled={busy}
-            className="rounded bg-gray-900 px-3 py-1.5 text-sm text-white disabled:opacity-50 dark:bg-gray-100 dark:text-gray-900"
-          >
-            {busy ? t("starting") : t("retranscribeTitle")}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function SegmentEditModal({
-  seg,
-  onClose,
-  onSave,
-}: {
-  seg: SegmentDto;
-  onClose: () => void;
-  onSave: (text: string | null) => Promise<void>;
-}) {
-  const { t } = useTranslation("workspace");
-  const [text, setText] = useState(seg.text);
-  const [busy, setBusy] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const taRef = useRef<HTMLTextAreaElement>(null);
-  const revised = seg.revised != null;
-
-  // Grow the textarea to fit its content; CSS max-height keeps the modal on-screen (it scrolls past that).
-  function autosize() {
-    const el = taRef.current;
-    if (!el) return;
-    el.style.height = "auto";
-    el.style.height = `${el.scrollHeight}px`;
-  }
-  // Size to the initial text once mounted.
-  useEffect(() => autosize(), []);
-
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => e.key === "Escape" && onClose();
-    document.addEventListener("keydown", onKey);
-    return () => document.removeEventListener("keydown", onKey);
-  }, [onClose]);
-
-  async function save(value: string | null) {
-    setBusy(true);
-    setError(null);
-    try {
-      await onSave(value);
-    } catch (e) {
-      setError(apiErrorMessage(e, t("errSaveSegment")));
-      setBusy(false);
-    }
-  }
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/40 p-4" onClick={onClose}>
-      <div
-        role="dialog"
-        aria-label={t("editSegment")}
-        className="w-full max-w-3xl rounded-lg border bg-white p-5 shadow-xl dark:border-gray-700 dark:bg-gray-900"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <h2 className="mb-3 text-base font-semibold dark:text-gray-100">{t("editSegment")}</h2>
-        <textarea
-          ref={taRef}
-          autoFocus
-          value={text}
-          onChange={(e) => {
-            setText(e.target.value);
-            autosize();
-          }}
-          aria-label={t("segmentTextAria")}
-          className="block max-h-[60vh] min-h-[8rem] w-full resize-none overflow-y-auto rounded border px-2 py-1 text-sm dark:border-gray-700 dark:bg-gray-800 dark:text-gray-100"
-        />
-        {revised && (
-          <p className="mt-2 text-xs text-gray-500 dark:text-gray-400">
-            <span className="font-medium">{t("originalLabel")}</span> {seg.original}
-          </p>
-        )}
-        {error && <p className="mt-2 text-sm text-red-600 dark:text-red-400">{error}</p>}
-        <div className="mt-3 flex items-center justify-end gap-2">
-          {/* Clearing a revision restores the model's original words (sends null to the server). */}
-          {revised && (
-            <button
-              type="button"
-              onClick={() => save(null)}
-              disabled={busy}
-              className="mr-auto rounded border px-3 py-1.5 text-sm hover:bg-gray-50 disabled:opacity-50 dark:border-gray-700 dark:text-gray-200 dark:hover:bg-gray-800"
-            >
-              {t("resetToOriginal")}
-            </button>
-          )}
-          <button
-            type="button"
-            onClick={onClose}
-            className="rounded border px-3 py-1.5 text-sm hover:bg-gray-50 dark:border-gray-700 dark:text-gray-200 dark:hover:bg-gray-800"
-          >
-            {t("common:cancel")}
-          </button>
-          <button
-            type="button"
-            onClick={() => save(text)}
-            disabled={busy}
-            className="rounded bg-gray-900 px-3 py-1.5 text-sm text-white disabled:opacity-50 dark:bg-gray-100 dark:text-gray-900"
-          >
-            {busy ? t("common:saving") : t("common:save")}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function RecordingNameForm({
-  initial,
-  onSave,
-  onCancel,
-}: {
-  initial: string;
-  onSave: (name: string) => void;
-  onCancel: () => void;
-}) {
-  const { t } = useTranslation("workspace");
-  const [value, setValue] = useState(initial);
-  return (
-    <form
-      className="flex items-center gap-2"
-      onSubmit={(e) => {
-        e.preventDefault();
-        onSave(value);
-      }}
-    >
-      <input
-        autoFocus
-        value={value}
-        onChange={(e) => setValue(e.target.value)}
-        onKeyDown={(e) => e.key === "Escape" && onCancel()}
-        placeholder={t("recordingNamePlaceholder")}
-        aria-label={t("recordingNamePlaceholder")}
-        className="w-64 rounded border px-2 py-1 text-base dark:border-gray-700 dark:bg-gray-800 dark:text-gray-100"
-      />
-      <button type="submit" className="rounded border px-2 py-1 text-sm hover:bg-gray-50 dark:border-gray-700 dark:text-gray-200 dark:hover:bg-gray-800">
-        {t("common:save")}
-      </button>
-    </form>
-  );
-}
-
-export function SpeakerRow({
-  label,
-  info,
-  initial,
-  count,
-  durationMs,
-  canPlay,
-  playing,
-  selected,
-  onSelect,
-  onTogglePlay,
-  onDelete,
-  onAssign,
-  onCreate,
-  onMulti,
-  canManagePeople,
-  onEditPerson,
-}: {
-  label: string;
-  info: SpeakerInfo | undefined;
-  initial: string;
-  count: number;
-  durationMs: number;
-  canPlay: boolean;
-  playing: boolean;
-  selected: boolean;
-  onSelect: () => void;
-  onTogglePlay: () => void;
-  onDelete: (name: string) => void;
-  onAssign: (profileId: string | null) => void;
-  onCreate: (name: string) => void;
-  onMulti: () => void;
-  canManagePeople: boolean;
-  onEditPerson: () => void;
-}) {
-  const { t } = useTranslation("workspace");
-  // The display name for the per-speaker action labels (the assignment typeahead owns the editing UI).
-  const name = initial;
-  // The interactive controls (assign box, play, delete) sit inside the clickable row, so they stop event
-  // propagation to avoid toggling the row's selection when used.
-  const stop = (e: { stopPropagation: () => void }) => e.stopPropagation();
-
-  // One line per speaker: [label · auto] · assign typeahead · segment count · toolbar (play/delete). The whole
-  // row is a button that toggles the speaker's segment table below the list; the leading label column is
-  // fixed-width so the assign box and the items after it line up across rows.
-  return (
-    <div
-      role="button"
-      tabIndex={0}
-      aria-pressed={selected}
-      aria-label={t("selectSpeakerAria", { name })}
-      onClick={onSelect}
-      onKeyDown={(e) => {
-        if (e.key === "Enter" || e.key === " ") {
-          e.preventDefault();
-          onSelect();
-        }
-      }}
-      className={`flex cursor-pointer flex-wrap items-center gap-2 rounded-lg border px-2 py-1 hover:bg-gray-50 dark:hover:bg-gray-800 ${
-        selected
-          ? "border-blue-400 bg-blue-50 ring-1 ring-blue-300 dark:border-blue-600 dark:bg-blue-900/30 dark:ring-blue-700"
-          : "border-transparent"
-      }`}
-    >
-      <div className="flex w-32 shrink-0 items-center gap-1">
-        <span className="text-xs text-gray-400 dark:text-gray-500">{label}</span>
-        {info?.identifiedAuto && (
-          <span
-            title={t("autoNameTitle")}
-            className="rounded bg-blue-100 px-1 text-[10px] font-medium text-blue-700 dark:bg-blue-900/40 dark:text-blue-300"
-          >
-            {t("autoBadge")}
-          </span>
-        )}
-      </div>
-      <div onClick={stop} onKeyDown={stop}>
-        <SpeakerAssign
-          label={label}
-          // The trigger reads its current label from here. Without it every identified speaker showed
-          // "Unassigned" while the row's own badges said otherwise.
-          //
-          // Only when it differs from the raw label: this row already prints SPEAKER_nn in the column to
-          // the left, so echoing it in the dropdown would say nothing. The transcript row passes its name
-          // unconditionally on purpose - there the trigger replaces the speaker label itself.
-          displayName={info && info.displayName !== label ? info.displayName : undefined}
-          isMulti={info?.isMultiSpeaker ?? false}
-          onAssign={onAssign}
-          onCreate={onCreate}
-          onMulti={onMulti}
-        />
-      </div>
-      {/* Who this person is, when we know. Deliberately absent for an anonymous speaker and for a
-          "Multiple Speakers" slot - the server sends null for both, and inventing something here would
-          claim more than Diariz knows. */}
-      {info?.personId && !info.isMultiSpeaker && (info.title || info.companyName || info.isInternal !== null) && (
-        <span className="flex min-w-0 shrink items-center gap-2 text-xs text-gray-500 dark:text-gray-400">
-          {(info.title || info.companyName) && (
-            <span className="truncate">
-              {info.title && info.companyName
-                ? t("speakerAt", { title: info.title, company: info.companyName })
-                : (info.title ?? info.companyName)}
-            </span>
-          )}
-          {info.isInternal !== null && (
-            <span
-              // The row has room for a job title and a company and no more, so the rest of the person hangs
-              // off the chip. Same text as the contact card below, from one function, so they cannot drift.
-              title={contactSummary(info, {
-                internal: t("speakerInternal"),
-                external: t("speakerExternal"),
-                none: t("speakerNoContactDetails"),
-              })}
-              className={`shrink-0 cursor-help rounded px-1 text-[10px] font-medium ${
-                info.isInternal
-                  ? "bg-gray-100 text-gray-600 dark:bg-gray-800 dark:text-gray-300"
-                  : "bg-amber-100 text-amber-800 dark:bg-amber-900/40 dark:text-amber-300"
-              }`}
-            >
-              {info.isInternal ? t("speakerInternal") : t("speakerExternal")}
-            </span>
-          )}
-        </span>
-      )}
-      <span className="w-40 shrink-0 whitespace-nowrap text-xs text-gray-500 dark:text-gray-400">
-        {t("speakerSegmentCount", { count })} · {formatDuration(durationMs)}
-      </span>
-      <div className="flex items-center gap-0.5" onClick={stop} onKeyDown={stop}>
-        {/* Correcting who someone is belongs where you noticed they were wrong. Only for a real person - an
-            anonymous speaker and a multi-speaker slot have no record to edit - and only with the permission
-            the endpoint behind it enforces. */}
-        {canManagePeople && info?.personId && !info.isMultiSpeaker && (
-          <ToolbarButton label={t("editPersonAction", { name })} icon={PencilIcon} onClick={onEditPerson} />
-        )}
-        {canPlay && (
-          <ToolbarButton
-            label={playing ? t("pauseSpeaker") : t("playSpeaker", { label: name })}
-            icon={playing ? PauseIcon : PlayIcon}
-            active={playing}
-            onClick={onTogglePlay}
-          />
-        )}
-        <ToolbarButton label={t("deleteSpeaker", { label: name })} icon={TrashIcon} onClick={() => onDelete(name)} />
-      </div>
-    </div>
-  );
-}
-
-/// A note-taker's note woven into the transcript: same row layout as a segment (timestamp · speaker · text)
-/// but the current user is the "speaker" and the text is green, to distinguish it from transcribed speech.
-function NoteRow({ note, speaker }: { note: MeetingNote; speaker: string }) {
-  return (
-    <li className="flex items-start gap-3 rounded-lg border border-green-200 bg-green-50/40 px-4 py-2 dark:border-green-900 dark:bg-green-950/20">
-      <span className="w-12 shrink-0 font-mono text-xs text-gray-400 dark:text-gray-500">
-        {note.capturedAtMs != null ? fmt(note.capturedAtMs) : ""}
-      </span>
-      <span className="w-28 shrink-0 text-sm font-medium text-green-700 dark:text-green-400">{speaker}</span>
-      <span className="flex-1 whitespace-pre-wrap break-words text-sm text-green-700 dark:text-green-400">{note.text}</span>
-    </li>
-  );
-}
-
-/// What a transcript row needs to offer the Speakers-tab assignment typeahead in its speaker column, so a
-/// speaker can be named while reading/playing the transcript. Omitted (Speakers tab, where the speaker row
-/// above already carries the typeahead) → the row shows a plain label as before.
-export type SegmentAssign = {
-  infoOf: (label: string) => SpeakerInfo | undefined;
-  onAssign: (label: string, profileId: string | null) => void | Promise<void>;
-  onCreate: (label: string, name: string) => void | Promise<void>;
-  onMulti: (label: string) => void | Promise<void>;
-};
-
-function SegmentRow({
-  seg,
-  speakerName,
-  assign,
-  active,
-  selected,
-  selectMode,
-  showOriginal,
-  onClick,
-}: {
-  seg: SegmentDto;
-  /// The speaker name to show (localised "Multiple Speakers" overrides the server display).
-  speakerName: string;
-  assign?: SegmentAssign;
-  /// Currently playing (highlighted by the audio position).
-  active: boolean;
-  /// Picked in the transcript selection (drives the toolbar's bulk actions).
-  selected: boolean;
-  selectMode: boolean;
-  showOriginal: boolean;
-  /// Click anywhere on the row: select it (single, or toggle in Select mode). No longer auto-plays.
-  onClick: () => void;
-}) {
-  const { t } = useTranslation("workspace");
-  const revised = seg.revised != null;
-  return (
-    <li
-      id={`seg-${seg.id}`}
-      onClick={onClick}
-      className={`flex cursor-pointer items-start gap-3 rounded-lg border px-4 py-2 hover:bg-gray-50 dark:hover:bg-gray-800 ${
-        selected
-          ? "border-blue-400 bg-blue-50 ring-1 ring-blue-300 dark:border-blue-600 dark:bg-blue-900/30 dark:ring-blue-700"
-          : active
-            ? "border-blue-300 bg-blue-50 dark:border-blue-700 dark:bg-blue-900/30"
-            : "border-gray-200 bg-white dark:border-gray-700 dark:bg-gray-900"
-      }`}
-    >
-      {selectMode && (
-        // Visual only — the whole row is the click target, which toggles selection.
-        <input type="checkbox" checked={selected} readOnly tabIndex={-1} aria-hidden className="mt-1 shrink-0 pointer-events-none" />
-      )}
-      <span className="w-12 shrink-0 font-mono text-xs text-gray-400 dark:text-gray-500">{fmt(seg.startMs)}</span>
-      {assign ? (
-        // The typeahead is a click target inside a clickable row, so it stops propagation (which would
-        // otherwise select the segment). Closed, it renders as one button - cheap enough per row.
-        <div className="shrink-0" onClick={(e) => e.stopPropagation()}>
-          <SpeakerAssign
-            label={seg.speaker}
-            width="w-40"
-            subtle
-            displayName={speakerName}
-            isMulti={assign.infoOf(seg.speaker)?.isMultiSpeaker ?? false}
-            onAssign={(profileId) => assign.onAssign(seg.speaker, profileId)}
-            onCreate={(name) => assign.onCreate(seg.speaker, name)}
-            onMulti={() => assign.onMulti(seg.speaker)}
-          />
-        </div>
-      ) : (
-        <span className="w-28 shrink-0 text-sm font-medium text-gray-700 dark:text-gray-200">{speakerName}</span>
-      )}
-      {/* Auto-expands vertically to show the full (possibly merged) block of text. */}
-      <span className="flex-1 whitespace-pre-wrap break-words text-sm dark:text-gray-200">
-        {segmentText(seg, showOriginal)}
-      </span>
-      {/* The currently-playing row gets a small ▶ marker (distinct from the selection highlight). */}
-      {active && <span aria-hidden className="mt-0.5 shrink-0 text-blue-500 dark:text-blue-400">▶</span>}
-      {/* Marks a segment whose text has been edited or translated (a revision exists). */}
-      {revised && (
-        <span
-          title={t("revisedTitle")}
-          aria-label={t("editedAria")}
-          className="mt-0.5 shrink-0 text-teal-500 dark:text-teal-400"
-        >
-          ✎
-        </span>
-      )}
-    </li>
   );
 }
